requirements.txt
# RAG系统依赖包

# 核心框架
llama-index>=0.10.0
llama-index-embeddings-huggingface>=0.1.0  # HuggingFace embedding支持
llama-index-embeddings-openai>=0.1.0  # OpenAI embedding支持（可选）
llama-index-vector-stores-chroma>=0.1.0  # ChromaDB向量存储支持
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
sentence-transformers>=5.1.2
# 向量数据库
chromadb>=0.4.0

# 文档处理
reportlab>=4.0.0  # PDF生成
python-docx>=1.1.0  # Word文档处理
pypdf>=3.0.0  # PDF读取
docx2txt>=0.9
# 模型支持
openai>=1.0.0  # OpenAI API
dashscope>=1.17.0  # 通义千问API

# 中文embedding模型支持
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# 工具库
pyyaml>=6.0
python-dotenv>=1.0.0

# Web框架（用于API服务）
flask>=2.3.0
flask-cors>=4.0.0

# 可选：如果需要使用其他embedding模型
# huggingface-hub>=0.16.0

